[bbagent] Handle unregistgered sub-build proto stream
When launching a merge step, it is possible that the parent step record
a `$build.proto` log first and then invoke the child luciexe (i.e. the
build.proto stream of child luciexe may be registered later). The current
implementation assumes that the child build.proto stream will always be
present when merging step and will result in nil pointer if not so.

R=iannucci

Bug: 1017831
Change-Id: Ibefadbc58e419717d69787a72c2549cce67ed4eb