[bbagent] Handle invalid sub-build proto stream (e.g. non-existent or incomplete)
When launch a merge step, it is possible that the parent step record a
`$build.proto` log first and then invoke the child luciexe (i.e. the
build.proto stream of child luciexe may be registered later). The current
implementation assumes that the child build.proto stream will always be
present when merging step and will result in nil pointer if not so.

This change fixes the aforementioned issue by providing a default
INFRA_FAILURE build if no valid sub-build can be found at the moment.
If the child build proto comes in later, merge will happen as normal.
If the child build proto is never registered, the final step status will
be INFRA_FAILURE.

R=iannucci

Bug: 1017831
Change-Id: Ibefadbc58e419717d69787a72c2549cce67ed4eb