Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> imo just drop the full url stuff entirely

Oh.. I thought I document it somewhere but apparently I forgot that. I also want to get rid of this  functionality entirely but it looks like kitchen is still depending on it[1]. So we have to support both ways right now. I'll add a todo for cleaning this up after we fully deprecate kitchen.

[1]: https://source.chromium.org/chromium/infra/infra/+/master:go/src/infra/tools/kitchen/update_build.go;l=186;drc=dd9d48bd236d1abcf276b70fe9749001cafdae1d

Patch-set: 2
Label: Commit-Queue=0