Add logging for mmutex exclusive workflow
Currently when using mmutex to invoke puppet, there no log for the mmutex workflow. It will be difficult when debug issues like crbug.com/1038136.
This CL adds some logs for the exclusive workflow, targeting on the drain file operations. The shared mmutex workflow is considered as well. (I'm not sure who is using the shared execution though.)

Bug: chromium:1038136
Change-Id: I10d4df6f57b798c07f8a1f5332218cc75bbc3d06