Update patch set 3
Patch Set 3: -Auto-Submit -Commit-Queue

> Patch Set 2: Code-Review+1
> 
> test?

It is actually covered in this test[1] already. The test proto[2] simulates the structure of batch request really well. In the latest patchset, I've also included the change for regenerating this proto. I can confirm that, after regenerating the proto, without this fix, this test fails with the exact same error message mentioned in the bug.


[1]: https://source.chromium.org/chromium/infra/infra/+/master:go/src/go.chromium.org/luci/common/proto/fieldmasks_test.go;l=97-120;drc=b62ad81ae4fc8919e41a987ba5ca15cb96831c5c
[2]: https://source.chromium.org/chromium/infra/infra/+/master:go/src/go.chromium.org/luci/common/proto/internal/testingpb/testing.proto;l=29-37;drc=eb44c96b192ec098cbac949fd4a5d50d45d41cc0

Patch-set: 3
Label: Commit-Queue=0
Label: Auto-Submit=0