[cfgcache] Attempt to populate the cache on a cache miss.
Do it very carefully to avoid bombarding LUCI Config with tons of
requests if the datastore cache is cold. In the worst case it will
be receiving O(<rate of new process spawns>) QPS which should be
tolerable.

This will simplify deployment of changes that add new cfgcache.
Without this mechanism each such change will have to be deployed in
two stages:
  1. Deploy code that calls Update(...), so it warms the cache.
  2. Deploy code that calls Fetch(...) or Get(...).

Now its fine to deploy everything at once if its OK to have a
slightly lower availability before the first Update call.

R=iannucci@chromium.org, ddoman@chromium.org
BUG=1090916

Change-Id: I250d2de938ea286e1083127bfd66b98c56bd4e28