Update patch set 18
Patch Set 18:

Haven't read the whole CL, just the description.

Using Redis Set datatype makes sense only if it is built incrementally (or via server-side commands like SUNION) or queried via SISMEMBER. If all you do is **always** writing and reading the set in its entirety, it would be better to use regular Redis GET and SET and represent the set as a protobuf or something on the client side. It will cost Redis significantly less CPU (and it matters because there's only one Redis server, it won't scale horizontally).

Patch-set: 18