Update patch set 2
Patch Set 2:

(2 comments)

I think there will be people that didn't see your announcement (it is easy not to notice emails), but will notice this banner (it is hard not to nice it). Such people might not understand what is going on. How about adding a link to a GDoc/Markdown doc explaining this.

Also I think we should land this after we are certainly ready:
- Do we already display exonerations in UI?
- +Chan is fixing some things now, so we'd need Chan LGTM on this CL.

Patch-set: 2
Reviewer: Gerrit User 1157909 <1157909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>