Update patch set 3
Patch Set 3:

(1 comment)

Code looks fine.

Since this header is sort of unpleasant and we offer the user no way to get rid of it (there's not even a way to opt in to the result UI by default yet), should we consider only showing it if the build failed?

Or maybe we could consider a slightly less disruptive but still eye-catching button a little lower down (i.e. not something that shifts the whole page). Thoughts?

IMO a banner like this is best for when we want to push people into opting in to something, but we don't have that option yet.

Patch-set: 3