Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 2:
> 
> (2 comments)
> 
> I think there will be people that didn't see your announcement (it is easy not to notice emails), but will notice this banner (it is hard not to nice it). Such people might not understand what is going on. How about adding a link to a GDoc/Markdown doc explaining this.
> 
> Also I think we should land this after we are certainly ready:
> - Do we already display exonerations in UI?
> - +Chan is fixing some things now, so we'd need Chan LGTM on this CL.

Nodir, to clarify, the fixes you're referring to are the ones for discrepancies between resultdb and builds, right? I think the fixes are for a small amount of builds, so I'm not sure if they should block this change: it's mainly adding a banner for people to try the new UI, we are not changing the UI entirely yet.

Weiwei, there are trailing spaces in many html files, please fix.

Patch-set: 3