Update patch set 1
Patch Set 1:

Background:

Feature flags: led has multiple compilation modes, and I'd like to be able to distinguish between them when users give me reports.

The JSON output is possibly a bit half-baked, but what I'd like to do is have our builder scripts be able to (optionally) set symver tags and refs when uploading packages.

I'm thinking that if the option is set in the yaml, then when the builder uploads the package, it would:
  * run the tool(1) to extract its version info (say, 1.2.3)
  * Check to see if there's already a tag 1.2.3. If not, then:
  * Set the ref 1.x, 1.2.x to the new package, and set the tag `version:1.2.3`

This would (hopefully) have the effect that bumping the symver in the source is enough to get a new
set of refs/pins updated, and would allow (for example) recipes to depend on '1.x', rather than 'this hyper-precise pin which needs to be touched constantly' or 'latest'. The understanding would be that depending on these refs is mostly-but-not-entirely stable.

In this scheme, if someone updates the code without bumping the symver, nothing happens.

(1) This gets tricky when cross-compiling; I was thinking that for packages like this I would generate a 'bonus' copy of the binary to interrogate, though this gets awkward if someone wants to switch on runtime.GOOS or something.

Patch-set: 1
Label: Commit-Queue=0