Support arbtirary rotation URLs in LUCI-Notify.
The existing field wasn't used, so we just remove it, but we reserve the
ID just in case. Initially we planned to have two different fields, one
for Rota-NG and one for internal rotations. As it turns out, we can
access both via HTTP endpoints exposing JSON in the same format, so we
can just have a generic field containing a URL instead of
service-specific fields.

Bug: 1095871
Change-Id: Iadc10976c54aa07f130e4ee3e40a66cbaad7c9b8