[resultdb] Optimize invocations.Reachable.
# Problem

Current implementation of invocations.Reachable is naive: for each node
being visited, it reads node's outgoing edges from Spanner.
This is slow for two reasons:
- A lot of Spanner RPCs, e.g. 1077 invocations are reachable from
  build 8877501336285858192 => 1077 RPCs.
- The RPCs parallelism is limited: we have to fetch outgoing edges
  before going deeper.

# Solution

Add invocations.ReachCache type which can read/write a cached set of
all invocations reachable from a given one, stored in Redis.
The set is guaranteed to be either correct or absent.

Read ReachCache when reading reachable invocations: when visiting a
node, first check the cache:
- on cache-miss, continue the recursion
- on cache-hit, add the cached set to the overall set and
  skip visiting children (stops recursion).

Write ReachCache only when we have a guarantee that the set of reachable
invocations is immutable: when finalizing an invocation
and when creating a finalized invocation.

This design is chosen to preserve the current strong consistency
guarantees. It is resilient to a failure of Redis and to
a lost of acked writes due to failover to the secondary replica,
see https://cloud.google.com/memorystore/docs/redis/high-availability

This design is also based on the current user behavior: normally users
do not read an invocation until they receive a notification about
build/cq_attempt completion. Thus, by the time they view the results,
most of the nodes are finalized.

# Observations

The following (anecdotal) observations were made. The subject
was build 8877501336285858192 with 1077 directly/indirectly
included invocations.

Without any caching (before this CL), invocations.Reachable() took
416ms. It does a Spanner.StreamingRead for each of 1078 invocations,
including leafs.

With task-level caching (before the build invocation is finalized),
invocations.Reachable() took 77ms: 36ms was spent on
fetching build's 178 outgoing edges. The rest was spent on 178
concurrent Redis reads, 10-30ms each.
Note of some of these 10-30ms was spent on opening a new Redis
connection.

With build-level caching (after the build is finalized),
invocations.Reachable() took 4ms and returned all 1077 included
invocations.

Bug: 1090197
Change-Id: Ie1feff6406daf3ed9e90894d00b3e191f2cb9550