Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is a path of pain and suffering, since it has a LOT of assumptions that no longer hold on GAEv2 or which make no sense. Better not to touch appengine/gaemiddleware. It is something we want to bury with GAEv1.
> 
> See this example instead:
> https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/master/examples/appengine/helloworld_v2/

Hi Vadim,

Could you elaborate on the problems with gaemiddleware?

I'm not sure if you are including the abstraction layer, appengine/impl/{prod,cloud} under the things you want to deprecate. These seem incredibly useful in terms of not having to change all of the call sites for a migration such as this one. Are you advising that we could use these abstractions without gaemiddleware?

Patch-set: 1