Update patch set 1
Patch Set 1:

> > Is it still possible to toggle between classic and cloud datastore, up until we actually move to the new runtime?
> 
> Dynamically? No. But you can spin up a parallel GAE go1.12 module and route a percent of the traffic there. Most changes will be in main.go to initialize the server environment and routes through server.Server instead of gaemiddleware and appengine.Main(). 

Not dynamically, but I was hoping to be able to toggle it via an environment variable at startup, as I did in main.go in this change. That way we can roll out the cloud ds change ahead of and independently from the runtime version change, which should reduce the risk at each step.

I'll poke around a bit and see how this would work with luci/server.

> 
> You may need to disable dscache ("Disable datastore cache" https://luci-notify.appspot.com/admin/portal/gae) while running in this mode since GAE's memcache is not available from GAEv2 so if datastore is written from GAEv2, it can't invalidate GAEv1's cache entries. This applies even if you proceed with hybrid GAEv1-Cloud context approach.

Do you mean specifically if we run classic datastore and cloud datastore in parallel, we'd want this so the classic instances don't rely on cache consistency? (It doesn't seem like it would matter if we're only running one version or the other)

Patch-set: 1