[resultdb] Add test_location tag
test_location tag has format "filename:lineno" where ":lineno" part is
optional. It points to the location where the test is defined.
Generate the tag from GTest and JSON Test Results.

Strictly speaking, the filename is not fully-qualified because it does
not have repo URL, but this is just a tag (not a proto field). We
can change it later.

This will be used to filter based on the source code path.

Change-Id: Id1c8aa20f1257614a782b12574cb1e93ac386bb8