Update patch set 3
Patch Set 3:

Hmm, some tests don't actually build executables to be run out of the build directory and instead invoke the script directly from the source tree, and that's probably what's causing this case to happen. That's probably ok but it's worth thinking if that will affect other systems.

Patch-set: 3