Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> this is at least ambiguous, twice:
> 1) there may be multiple with-code errors in a multi-error
> 2) code may legitimately attach an Unknown error, and it would be ignored
> 
> can we keep this ad-hoc logic in Milo?

I can certainly move this to Milo, but doesn't it make sense to fix it at the source? Otherwise other apps using this library may run into the same error.

The question is, what should the behaviour of grpcutil.Code be for a multi-error? Are you arguing that it should always return Unknown? If so, we should document that and enforce it in the tests.

Patch-set: 1
Label: Commit-Queue=0