[Resultdb] Fail fast if fail to parse a known JTR result using formats.JSONTestResults
If locationPrefix is not empty, it should mean the result to be parsed has json test result format. So in the case if we fail to parse the result, no need to try gtest parser, instead we can return the error right away. We should not panic in this case though, since we now know it could happen.

Bug: 1102709
Change-Id: I4d2ac04192f175c1392fce17b29d24d5746320af
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2284467
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>