[lucicfg] Refactor realms impl to reuse it for internal realms.
For a variety of reasons reusing luci.realm(...) etc. directly from
infradata/config Starlark code is fragile. Instead totally divorce
the logic of defining realms from the rest of default lucicfg model,
and just "instantiate" it in the stdlib and, separately, in internal
configs, using different keys. That way they never conflict.

The only logical change is that luci.realm(...) definition is
idempotent now. It simplifies internal configs a bit.

R=tandrii@chromium.org, iannucci@chromium.org
BUG=1051724

Change-Id: I9691b33f80904139b191d91f16214e5c3be147a0