[isolate] Add cas.ChunkerDeduper in prep for supporting batch archive
For the chunkers produced from each `.isolate`, we pass them into this deduper so that we will only upload chunkers that haven't been requested to be uploaded in the client.

Please let me know if you think this helper looks OK. If so, I will go add some tests for it.

Bug: 924625
Change-Id: If6f115f3dd0abc6c3daf180296a9f4cac8c1edc5