Fix the bot username in LUCI-Notify tree closing.
This is the username we look for when checking if the last update to the
tree was made by a human. If a different user closed the tree we assume
it's a human and leave the tree alone, to avoid overwriting closures
made for reasons beyond our ken.

The previous username is just wrong, so we'd detect our own automatic
tree closures as manual closures, and so never automatically re-open the
tree.

Bug: 1094768,1106207
Change-Id: Ia90d8d9e698e29f553c4b7532c26d19403546d9d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2305493
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>