[buildbucket] Fix GetIgnoreMissing.
It didn't work correctly when:
  1. `dst` contains only one item (in particular a slice).
  2. `dst` contains multiple slices.

Case (1) happens when fetching a slice of TagIndex shards.

In case (1) datastore.Get may return ErrNoSuchEntity directly,
not wrapped in a MultiError. In case (2) datastore.Get may return
a multi-error of multi-errors.

errors.Filter can deal with all this cases already, so just reuse
it.

R=tandrii@chromium.org, smut@google.com

Change-Id: Iec96f8658132c00221022e43e3ec34292fb0787b