[buildbucket] Add a wrapper for identity.Identity and filter by it in SearchBuilds
The type of `created_by` in Datastore Build Entity is BlobProperty,
which is just the bytes of the  string form - <kind>:<name>.

To be compatible, this CL does:
1. Add query filter by createdBy in SearchBuilds
2. Create an Identity struct as a wrapper for identity.Identity and implements the datastore.PropertyConverter.

The reason why the datastore.PropertyConverter shouldn’t be directly added inside `auth/identity` pkg is that
it’s imported by many other packages. Some of them are not related to GAE.

The []byte stored by Python can be loaded as (utf-8) string in Go (Thanks Vadim to tell me about this).
But the reason I don’t directly replace []byte with string in Entity
is that there is an overlapping runtime of both Py and Go versions.
If some writing operations live in Go which write a string into created_by, while some reading operations still live in Python,
there will be some issues of reading the `created_by` string in Py.

Bug:1090540
Change-Id: Id22339ca94fe3039703c5618339299732fcfdc6e