[mmutex] Set the max interval for lock acquisition to one second
From the default settings in lock_file_helpers.go, the default timeout is two hours and the lock acquisition attempt count is 100. That means the default interval for file acquisition is 1.2 minutes, which is 72 seconds. This is too long and blocking itself when the first acquisition fails. In the case of the crbug.com/1038136, after swarming task finishes, it bot will reboot. Due to the long interval, mmutex does not come up to acquire the lock and delete the drain file, which will block the next swarming task.

This CL set the max interval to be one second.

Bug: chromium:1038136
Change-Id: I9cfde541ef5e4b52427cdddd5808e12099cc2aa7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2309214
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>