[resultdb] Parallelize test variant query
Spanner runs at most 50 concurrent sub-queries. This results in a very
poor performance for large queries. A discussion with the Cloud Spanner
team about increasing the limit is ongoing.
In the meantime, manually parallelize the Spanner queries used in
QueryTestResults RPC.

Anecdotally, this improves overall QueryTestResults RPC latency by 3x
for a graph size of 1000.

Bug: 1090197
Change-Id: I79a272e20918af264b7e9d4d82aeee2faad3cb40