[logdog] Fix butler to use byte-oriented named pipes on Windows.
The Microsoft/go-winio library has some special behavior when using
message-mode pipes. The documented claim is that:

  // MessageMode determines whether the pipe is in byte or message
  // mode. In either case the pipe is read in byte mode by default.
  // The only practical difference in this implementation is that
  // CloseWrite() is only supported for message mode pipes; CloseWrite()
  // is implemented as a zero-byte write, but zero-byte writes are only
  // transferred to the reader (and returned as io.EOF in this
  // implementation) when the pipe is in message mode.

I THOUGHT that there was an outstanding bug in butler which is
that the connection protocol can be racy. I thought that a client
could open the pipe, write the butler magic and prefix, write the
data, close the pipe and exit (causing butler teardown) before the
butler actually reads any data from it. See CL comments for why
I think this is no longer the case.

I wanted to prevent this by changing the protocol from:

   open, write magic, write data, close

To

   open, write magic, READ EMPTY MESSAGE, write data, close

Where the butler would only send the empty message once it had
actually registered the new stream.

However, this client side piece (the 'read') wasn't ever introduced...
and just as well because MessageMode=True has another bonus behavior!

If the Go server receives a zero-width write, it will return EOF...
And this EOF becomes sticky! Even though the pipe is not ACTUALLY
closed, and the client can write more data, from Go's PoV the
pipe is DONE.

Ok, so now, Powershell. I believe that PS uses zero-width messages
as delimiters between messages. This is because you can actually
interlink PS programs, even across the network, with NamedPipes
(in message mode). This message-oriented nature applies to
EVERYTHING in PowerShell, even when printing "help" (which is
ACTUALLY sending a string object to the Output stream).

So why did using an anonymous pipe "work"? Well, anonymous pipes are
always byte-mode, and zero-size writes are NOT communicated for
byte-mode pipes... at all. Even if PS writes such a message, the
recipe engine (the other end of the anonymous pipe) won't see it
(and thus, wouldn't ever write it to the logdog stream).

Finally... Why did cranking up the debug logs help? I'm honestly
not sure, but I suspect that there may be a race in the zero-width
message reading in Go; If the server is slow and the client is
interleaving "" messages with actual data, then I think the server
program may actually blur everything together and fly right past
the EOF. Maybe. The EOF detection is only tested (afaik) against
other Go namedpipe implementations, which means that they'll
cooperatively write "" and then stop sending data.

In conclusion: We can just use byte-mode named pipes and everything
works fine.

R=yiwei@chromium.org,tandrii@chromium.org

Bug: 1103028
Change-Id: I116c8f701ec1cd3e4b99715eb5c475dc5d2515c7