[resultdb] Skip reading cache in some cases
ReachCache is populated by finalizer service. Currently finalizer
indirectly reads the cache that is about to populate, which must be
spoiling the cache-hit stats because the cache is not there yet.

Skip reading the cache for the invocation that we are building cache
for.

Bug: 1112029
Change-Id: Ibcab2c536c7016bd1a4dd95a6536eb91623ba683
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2334322
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>