[resultdb] "With only unexpected results" filter
Add Expectancy.VARIANTS_WITH_ONLY_UNEXPECTED_RESULTS
which is similar to VARIANTS_WITH_UNEXPECTED_RESULTS, but ignores
variants that have a mix of expected and unexpected results.

It is implemented on top of VARIANTS_WITH_UNEXPECTED_RESULTS,
by grouping its result by TestID and VariantHash, excluding
those test variants that have mixed results, and returning the rest.
It does mean that the VARIANTS_WITH_UNEXPECTED_RESULTS query must
run completely, but it is no longer a problem with recent optimizations.

Bug: 1112127
Change-Id: I013ce88397420dd357857399c99292b169f22257