[server/tq] Decouple Scheduler from Dispatcher a bit more.
Instead of using Dispatcher as a tqtesting.Deserializer in a
Scheduler, pass deserialized protos explicitly. This works only
on a happy path, but we generally don't care about non-happy path
in server/tq client's tests.

This is a preparation for moving some of the Dispatcher's state
(in particular Submitter) into the context for easier replacement
in tests.

R=tandrii@chromium.org

Change-Id: I2c11a4229d333ea6489c18f5e2884eb78e570cdd