[luciexe/exe] Implement higher-level build sync/send functionality.
Previously, exe implementations were expected to coordinate wrangling
`build` and `send` throughout their entire program. This CL introduces
a (hopefully) better way of doing things.

This removes the 'send' function from MainFn, and makes `build` a
read-only value. The "live" build state is now embedded in the
context.

There are a couple of new functions to modify attributes of the
"live" Build. Invoking these gives exclusive read/write access
to that portion of the build.

Additionally, this adds WithStep to manage the lifecycle of Build
Steps, including:
  * Step namespacing (current namespace is embedded in context)
  * Step name deduplication
  * Step start/end times
  * Step status codes

This functionality also exposes SinkBuildUpdates, which would allow
users to wire up their own "sink" function for build protos, independent
of logdog. This could allow "luciexe" code to function in non-luciexe
contexts (by installing a nil or dummy sink into the context).

To be added in later CLs:
  * Automatically wiring of luci "logging" package to a "log"
    stream attached to the step.
  * Helper methods for making subprocess and Sub-Build steps.

R=pprabhu@google.com, yiwzhang@google.com, yulanlin@google.com

Change-Id: Ieaec022565143ab282a47627882e4a84acbbacbe