[resultdb] fix warmup timeout
The default value of ServerConfig.MaxWarmUpDuration is conditionally
set inside cfg.Validate(). However, the change is applied only to s.cfg,
not to the param `cfg` passed on Run().

Nevertheless, the timeout context for warmUp() was created with
the original cfg. Therefore, the default value is not correctly applied
to the timeout context.

This CL changes warmUp() as a member function of Server() so that it always
uses s.cfg.

R=chanli@chromium.org

Bug: 1113089
Change-Id: I1a1931e1786133f3d1332d7b6ced930c51d1f1fd