[WIP][BuildPage][Timeline] Load data into vis-timeline
I am not entirely convinced that using vis-timeline is the correct approach here. I think it violates some lit-element principles (like modifying the DOM outside of render). However, it seems to be the best library we can find, and is a direct replacement of the deprecated one we are currently using in Milo.

Another approach is to implement this ourselves.

Let me know what you think.


Bug:1115838
Change-Id: Ie03e221e291514560ffbe58a7fe1871a4124ed94