[server/tq] Wrap CreateTaskRequest in a reminder.Payload.
It will allow to add different kinds of task requests to it (in
particular pubsub.PublishRequest).

We wrap serialized protos, since *.proto with Cloud Tasks and
PubSub APIs are not available in GOPATH, and manually vendoring
them is complicated and fragile.

Other related changes:
1. tqpb.Extra is no longer need, it is part of reminder.Payload now.
2. Tracing headers are injected when CreateTaskRequest is created.
3. tq.Submitter is non-implementable externally now, but it should
   not be a problem, since no one should be needing to implement it
   (they should use tqtesting instead).

R=tandrii@chromium.org

Change-Id: Id89a2bddff1a883ada3e5aab023a9597e9214ab6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2359537
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>