[luciexe/exe/build] Add support for manipulating output properties.
Adds ability to manipulate the Output.Properties of the context-embedded
build.State.

This introduces ModifyProperties and NamespaceProperties.

NamespaceProperties allows higher levels of the stack to constrain
lower levels of the stack to subsections of the output properties
(otherwise the entire process has access to read and write the entire
output properties).

Change-Id: I6e00e73ce338cdc5f41c6fd7be7beda04b594969