[common/system/environ] Add general env-in-context construct.
This will be used for "recipes in Go" which need to do things like
manipulate $LOGDOG_NAMESPACE on a per-context basis. More generally,
it's frequently the case that existing (python) recipes have
need to manipulate the environment such that it's respected further
down the stack. Without a context-embedded environment I suspect
that recipes in Go will need to manipulate `os.Environ` which would
be Quite Bad(tm).

R=tandrii@chromium.org, vadimsh@chromium.org, yiwzhang@google.com

Change-Id: Ic239981d5098136b15a77a67de749b54755b5e91