[resultdb] Parallelize querying of indexed invocations.
In this change, each worker gets the reachable invocations and
then blocks on the worker streaming the previous invocation's
results, s.t. the results reachable from the same index point
(e.g. a timestamp or a commit position) remain together and can
later be internally sorted.

For this we use a channel per worker to signal that they are done
with their results.

R=nodir,chanli

Bug: 1074407
Change-Id: I3683ea6fb73603f3c64464ebae9fb8cfdd07eb85