Document cq retry quota config
I had to look through the CQ code to understand exactly how this works.
This is my attempt to document how it works now, as I understand it. I
can put this in our team docs if this isn't the appropriate place, but
it feels like the existing documentation doesn't really give you enough
information to know how to use the feature.

Change-Id: Ia99c14ca52649b95b724e67c49e373b38a50216f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2414755
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>