[rts] Fetch rejected patchsets with tests
Implement fetching of patchsets rejected due to test failures,
along with the tests that caused the failure.

The implementation is Chromium-specific, so keep the implementation
in the chromium package. Fetch the patchsets and test failures
using BigQuery, and cache them locally. Cache key is one day (in UTC),
and the time window is aligned with day boundaries.

On cold cache, it takes ~3min to execute the query and download ~300 MB
of data (for the window of 1w), but subsequent evaluations on the same
day are instantaneous.

Bug: 1112125
TBR=chanli@chromium.org
Change-Id: Iece46d6a1f9230fa6e0a740508c6619d9dc802a1