[rts] Compute precision
A precise RTS algorithm selects only affected tests.
Evaluate precision based on how many hours of testing would be
saved.

It is infeasible to fetch all durations of all tests that ran recently,
because of the data size. Instead fetch only a sample of durations,
up to 1M results from CQ in the past day.

To minimize the number of requests to Gerrit and select a random sample
at the same time, order results by hash of the patchset ID.

Bug: 1112125
Change-Id: I67955a6c88ee8296ee8558a42f00eb1be7cf032d