[rts] Add history package
The Browser Infra prefer to be able to fetch CQ rejections and other
historical data to a file, and then, when evaluating an RTS algorithm,
simply point to the file. The use case is to have a collection of
historical files representing different dates and builders.

Incidentally this simplifies the how domain-specific/domain-agnostic
boundary is expressed. Instead of using a Go-based Backend interface,
we can use files in a specific format.

This CL defines a file format and implements reader/writer.
The format is zstd-compressed recordio-encoded protobuf messages.
Each message is either a Rejection or a TestDuration, although
this specific CL does not define TestDuration.

Also generalize Rejection a bit by allowing multiple patchsets.

Next CLs will refactor the framework to read from a file instead,
and refactor chromium-specific code to be a binary that can save
historical files.

Bug: 1112125
Change-Id: I50f0572ab7549c261cd3be9dd77909f2289f14cf