[rts] Add variant support
Browser EngProd would like to try excluding certain test suites
from certain builders; and would like to try using rts framework for
evaluation of safety and efficiency.

Test suite and builder are expressed as test variants in ResultDB.
Add "Variant" to the Test proto and populate it based on ResultDB data.

This increases the amount of data fetched from BigQuery because
before a test was considered failed if any of its variants failed.

Bug: 1135389
Change-Id: Iffb4e13e3bf3bccd80a4451acd692313ad546780