cv: Add structs and more detailed todo notes in gobmap
This is from the CL partially implementing the interface
of gobmap, but with implementation removed for now.
Implementation was temporarily paused because we're waiting
for more design review to decide whether to use spanner,
and it would also be more convenient to implement after
the config package functions are landed.

Bug: 1123642
Change-Id: I0137c07b9313415a9bbaf6e60a74ea853060f2df