[rts] Batchify the RTS algorithm signature
Currently the RTS algorithm candidate accepts a single test as input,
while the evaluation code has the full set of failed tests.
This slows down the evaluation because the algorithm might have to
evaluate the same set of changed files multiple times unnecessarily.

Change the algorithm input to have multiple tests. The algorithm
returns whether it would run any of them. Currently there is no need
to know if the algorithm would run all of them.

R=tandrii@chromium.org

Bug: 1112125
Change-Id: I362343c5dbd87464ebd88e43ca51c1d5633c5556