[rts] Split rejection records in the history files
In practice, a bad patchset fails hundreds or thousands of tests.
Currently a historical Rejection record contains all failed tests,
which results in buffering the entire record before writing it to disk.

When logging how many records were written to a file, the logs are
misleading because the number of failed test variants might be 1000x
greater than the number of rejections.

Change the Record message to include a RejectionFragment, rather than
Rejection. A sequence of rejection fragments, ending with a terminal
one, constitutes one rejection.

Bug: 1112125
Change-Id: Iaa5d8a1ad8682bb61d1288fb0f107dc8451ed345