[rts] Improve Rejections query
Improve the code that excludes flaky tests.
Before this CL, a test variant could be considered flaky for one CL,
but not the other. This CL changes this, such that a test variant
is either flaky for the entire query, not at all.

A test variant is considered flaky if has mixed results
in N unique CLs. N=1 is too small because otherwise a single bad
patchset that introduces flakiness (but never lands), could make a
test flaky for everyone.

Do not exclude tests that reside in .js files

Exclude exonerated tests earlier in the query

Exclude skipped results, as they don't provide value for RTS purposes.
Change the time window.
Before this CL, we took wider window for CQ attempts.
=> sometimes a CQ attempt doesn't have all test results.
Instead, widen the time window for test results, such that we always
get all test results for a given CQ attempt.

Bug: 1112125
Change-Id: I913385b1fcc710a3ab020a4c29fb9320482f2484