[cipd] Add metadata events to BigQuery event log protos.
BQ rows can be up to 5MB in size, so we can potentially export
all metadata as is (since it is <512Kb). But it makes little sense
to export raw byte blobs. So we'll use metadata's content type
to decide if it should be exported or not (e.g. it might be useful
to export "text/plain" or "application/json" entries).

R=tandrii@chromium.org, yiwzhang@google.com
BUG=1141983

Change-Id: I408e820d1cddb3a52be1b7e0fe8ed4456fdd34ba