[filegraph] Implement "query" subcommand
Subcommand "query" returns reachable files in the distance-ascending
order (from closest to furthest).

Mostly this CL creates structure in the code, but doesn't implement
does not implement retrieval or traversal of the graph.
The Graph and Node are presented as interfaces, because their
implementation (data structure, retrieval) is git-specific,
while traversal is not.

Bug: 1136280
Change-Id: I3775ad46a3e31b34d62f415ae0ee499eea7ab9c5