[filegraph] Prellocate all edges
Preallocate all edges when reading a graph from a file.
It reduces the load time of chromium/src.git graph from 900-1000ms
to 800-900ms.

Also remove a TODO to consider writing edge only for nodes that have
them. It didn't help because all files have edges, and vast majority
of nodes are files.

Bug: 1136280
Change-Id: I26690370e2647f9a9cbd2717a279b1528777e03e