[filegraph] Update git.Graph semantics.
The initial implementation of filegraph kept the graph synchronized with
the git tree: a file in the graph exists if and only if the file exists
in the git tree of the most-recently-synchronized commit.
This means we can incrementally update the graph by following commits in
the chronological order. In particular, if a file was moved/removed
in the git commit, and it is moved/removed in the graph.
Here the update operation requires the state of the graph for the
parent commit.

However, this fails short on commits that have multiple parents.
So the initial implementation passed --first-parent to git-log, with
a TODO to fix it.

Stepping back, the graph was trying to represent two things at the same
time: 1) the state of the file tree and 2) aggregated changes to the
file tree. To support for non-first parents of a commit, this CL
removes (1), and keeps (2) which is what we need.
In particular, now a file in the graph is never removed/moved,
and the graph is now append-only.

Primarily, this means the update operation no longer requires the
state of the graph for the parent commit, and in fact no longer depends
on the order of commits. This means we no longer have to pass
--first-parent.

This does mean that the graph has nodes for no-longer existing files.
But it also has a nice property that the history is never lost.
For example, if a file was removed, but then the change was reverted,
this information is not lost when processing the removal.

Still, we need an alternative solution for file renames.
Introduce "alias edges", which are edges with distance 0.
Reuse edge struct for memory efficiency.

R=guterman@google.com, tandrii@chromium.org

Bug: 1136280
Change-Id: Id2cd775a9c61e4fe8375a8946812adf393ecb721